Create views folder to split up views into multiple files. Created a base_view module that shares common imports.
